refactor: IProjectTags coupled to apiHandler for better readability
