+++ client/src/features/project-list/components/ProjectCards.tsx
@@ -1,6 +1,7 @@
 import ADDRESSES from "../../../constants/addresses/addresses";
 import { useNavigate } from "react-router-dom";
-import { IProjectDetails, IProjectTags } from "@/types";
+import { IProjectDetails } from "@/types";
+import { IProjectTags } from "@/utils/data/ProjectDataService";
 
 interface ProjectCardProps {
   pageNumber: number,
